# src/RSCD/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "CSC Api",
    "description" : "This document contains the API descriprion",
    "contact" : {
      "name" : "Cloud Signature Consortium",
      "url" : "https://cloudsignatureconsortium.org/",
      "email" : "info@cloudsignatureconsortium.org"
    },
    "license" : {
      "name" : "Cloud Signature Consortium License",
      "url" : "https://cloudsignatureconsortium.org/"
    },
    "version" : "1.0.3.0"
  },
  "servers" : [ {
    "url" : "https://{host}:443/csc/v1/",
    "variables" : {
      "host" : {
        "default" : "service.domain.org"
      }
    }
  } ],
  "paths" : {
...
    "/signatures/signHash" : {
      "post" : {
        "tags" : [ "signatures" ],
        "summary" : "Calculate the remote digital signature of one or multiple hash values provided as an input.",
        "operationId" : "signaturesSignHash",
        "requestBody" : {
          "description" : "Sign hash request object",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/input-signatures-signhash"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Sign hash response object",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/output-signatures-signhash"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid_request"
          }
        },
        "security" : [ {
          "BearerAuth" : [ ]
        } ]
      }
    },
    "/signatures/timestamp" : {
      "post" : {
        "tags" : [ "signatures" ],
        "summary" : "Generate a time-stamp token for the input hash value.",
        "operationId" : "signaturesTimestamp",
        "requestBody" : {
          "description" : "Timestamp request object",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/input-signatures-timestamp"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Timestamp response object",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/output-signatures-timestamp"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid_request"
          }
        },
        "security" : [ {
          "BearerAuth" : [ ]
        } ]
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "clientData" : {
        "type" : "string",
        "description" : "Arbitrary data from the Signature Application"
      },
      "credentialID" : {
        "type" : "string",
        "description" : "The identifier associated to the credential"
      },
...
      "hash" : {
        "type" : "array",
        "description" : "One or more Base64-encoded hash values to be signed",
        "items" : {
          "type" : "string",
          "description" : "Single Base64-encoded hash value to be signed"
        }
      },
      "hashAlgo" : {
        "type" : "string",
        "description" : "Specifies the OID of the algorithm used to calculate the hash value(s)"
      },
      "hashTBS" : {
        "type" : "string",
        "description" : "The Base64-encoded hash value to be time stamped"
      },
...
      "nonce" : {
        "type" : "string",
        "description" : "Specifies a large random number with a high probability that it is generated by the Signature Application only once"
      },
...
      "SAD" : {
        "type" : "string",
        "description" : "The Signature Activation Data to provide as input to the signatures/signHash method."
      },
...
      "signAlgo" : {
        "type" : "string",
        "description" : "Specifies the OID of the algorithm to use for signing"
      },
      "signAlgoParams" : {
        "type" : "string",
        "description" : "Specifies the Base64-encoded of DER-encoded ASN.1 signature parameters"
      },
      "signatures" : {
        "type" : "array",
        "description" : "Array of Base64-encoded signed hash",
        "items" : {
          "type" : "string",
          "description" : "One Base64-encoded signed hash"
        }
      },
      "timestamp" : {
        "type" : "string",
        "description" : "The Base64-encoded time-stamp token as defined in RFC 3161 as updated by RFC 5816"
      },
...
      "input-signatures-signhash" : {
        "required" : [ "SAD", "credentialID", "hash", "signAlgo" ],
        "type" : "object",
        "properties" : {
          "credentialID" : {
            "$ref" : "#/components/schemas/credentialID"
          },
          "SAD" : {
            "$ref" : "#/components/schemas/SAD"
          },
          "hash" : {
            "$ref" : "#/components/schemas/hash"
          },
          "hashAlgo" : {
            "$ref" : "#/components/schemas/hashAlgo"
          },
          "signAlgo" : {
            "$ref" : "#/components/schemas/signAlgo"
          },
          "signAlgoParams" : {
            "$ref" : "#/components/schemas/signAlgoParams"
          },
          "clientData" : {
            "$ref" : "#/components/schemas/clientData"
          }
        },
        "description" : "Input parameters for signatures/signHash method",
        "example" : {
          "SAD" : "_TiHRG-bAH3XlFQZ3ndFhkXf9P24/CKN69L8gdSYp5_pw",
          "hash" : [ "sTOgwOm+474gFj0q0x1iSNspKqbcse4IeiqlDg/HWuI=", "c1RPZ3dPbSs0NzRnRmowcTB4MWlTTnNwS3FiY3NlNEllaXFsRGcvSFd1ST0=" ],
          "hashAlgo" : "2.16.840.1.101.3.4.2.1",
          "signAlgo" : "1.2.840.113549.1.1.1",
          "clientData" : "12345678"
        }
      },
      "input-signatures-timestamp" : {
        "required" : [ "hash", "hashAlgo" ],
        "type" : "object",
        "properties" : {
          "hash" : {
            "$ref" : "#/components/schemas/hashTBS"
          },
          "hashAlgo" : {
            "$ref" : "#/components/schemas/hashAlgo"
          },
          "nonce" : {
            "$ref" : "#/components/schemas/nonce"
          },
          "clientData" : {
            "$ref" : "#/components/schemas/clientData"
          }
        },
        "description" : "Input parameters for signatures/timestamp method",
        "example" : {
          "hash" : "sTOgwOm+474gFj0q0x1iSNspKqbcse4IeiqlDg/HWuI=",
          "hashAlgo" : "2.16.840.1.101.3.4.2.1",
          "clientData" : "12345678"
        }
      },
...
      "output-signatures-signhash" : {
        "required" : [ "signatures" ],
        "type" : "object",
        "properties" : {
          "signatures" : {
            "$ref" : "#/components/schemas/signatures"
          }
        },
        "description" : "Output parameters for signatures/signHash method",
        "example" : {
          "signatures" : [ "KedJuTob5gtvYx9qM3k3gm7kbLBwVbEQRl26S2tmXjqNND7MRGtoew==", "Idhef7xzgtvYx9qM3k3gm7kbLBwVbE98239S2tm8hUh85KKsfdowel==" ]
        }
      },
      "output-signatures-timestamp" : {
        "required" : [ "timestamp" ],
        "type" : "object",
        "properties" : {
          "timestamp" : {
            "$ref" : "#/components/schemas/timestamp"
          }
        },
        "description" : "Output parameters for signatures/timestamp method",
        "example" : {
          "timestamp" : "MGwCAQEGCSsGAQQB7U8CATAxMA0GCWCGSAFlAwQCAQUABCCrCqnrjH0VxXyQQlfnFJRx1jjrviTs7/GjKghr2AmluQIIVs5D8OUB4p4YDzIwMTQxMTE5MTEzMjM5WjADAgEBAgkAnWn2SSIWlXk="
        }
      }
    },
    "securitySchemes" : {
      "BasicAuth" : {
        "type" : "http",
        "scheme" : "basic"
      },
      "BearerAuth" : {
        "type" : "http",
        "scheme" : "bearer"
      }
    }
  }
}